Create typography module
